extract HighlightSearchTerm to standalone component
